rename 'app' dir to 'redux' to be more clear
