+++ client/src/features/home/Home.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
-import { getUser, RootState } from 'app/store';
-import { useAppDispatch, useAppSelector } from 'app/hooks';
+import { getUser, RootState } from 'redux/store';
+import { useAppDispatch, useAppSelector } from 'redux/hooks';
 
 /**
  * Home page for logged in user, currently does not really include anything
